backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
pyserial==3.5
opencv-python==4.10.0.84
numpy==1.26.4
pypylon==1.10.1
Pillow==10.4.0